ball processing logic added

+++ src/main/java/com/dugout/dugoutcore/dto/BallDto.java
@@ -18,7 +18,7 @@ public class BallDto {
   UserDTO bowler;
   UserDTO wicketKeeper;
   Boolean isValid;
-  Long runs;
+  Integer runs;
   BallType type;
   Integer ballNumber;
   Integer batsmanRuns;
